added Data for Model's Messages, SmsLogs

+++ apps/core/priv/repo/migrations/20230915123709_create_messages.exs
@@ -5,7 +5,7 @@ defmodule Core.Repo.Migrations.CreateMessages do
     create table(:messages, primary_key: false) do
       add :id, :uuid, primary_key: true, default: fragment("uuid_generate_v4()"), read_after_writes: true
       add :id_external, :string, null: true
-      add :id_tax, :integer, null: true
+      add :id_tax, :bigint, null: true
       add :id_telegram, :string, null: true
       add :message_body, :string, null: false
       add :message_expired_at, :utc_datetime_usec, null: true
